On exception, simply do a debug log and 503 error rather than logging a traceback and internal server error
